langgraphProject/pyproject.toml
[project]
name = "langgraphproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-tavily>=0.2.16",
    "langgraph>=1.0.6",
    "langgraph-cli[inmem]>=0.4.11",
    "langsmith>=0.6.4",
    "python-dotenv>=1.2.1",
]
